rename: asset directory segmentation
