+++ src/features/core/components/checkbox.tsx
@@ -1,7 +1,6 @@
 import { forwardRef } from 'react';
 
-import CheckIcon from '../../../assets/check.svg?react';
-
+import CheckIcon from '@/assets/icons/check.svg?react';
 import { cn } from '@/features/core';
 
 export const Checkbox = forwardRef<
